Added some CSV support to Parser Moved a lot of Processing/Merge code for easier maintenance Parser now merges the 'cost' field correctly Parser Cost Logic now entirely encapsulated within a custom Type to simplify everything TryConvert now has a lazy mode for improved performance where simple type casting is enough Achievements/Criteria now can do a huge amount of automation from in-game AchievementDB and WagoDB * CriteriaUID is used where possible * Nested Criteria via CriteriaTree is now used to determine Criteria for Achievements * Achievements/Criteria now incorporate more DB data, and will remove the 'achievement_criteria' symlink if possible * Many more Criteria now use _quests and _npcs to move themselves into the respective group which grants the Criteria * Criteria with _npcs which fail to move will revert into a simple 'crs' list instead * Criteria (which are not ignored by Blizzard) with available Item data will be linked with an Item provider (if the Criteria is not being incorporated, then the provider will be applied to the Achievement) * Achievements will only incorporate DB data when nested under a 'header' group or another Achievement * Achievements with a single Criteria will not incorporate that Criteria * Unsorted groups (i.e. /att unsorted) will not incorporate DB data * Known Issue: Instance Difficulty is not properly accounted for when moving Criteria for instance-related encounters
