+++ .contrib/Source Code/Parser/FieldTypes/IExportableField.cs
@@ -1,0 +1,7 @@
+﻿namespace ATT.FieldTypes
+{
+    public interface IExportableField
+    {
+        object AsExportType();
+    }
+}
