+++ .contrib/Source Code/Parser/Framework/ParseStage.cs
@@ -9,10 +9,12 @@
         NotStarted = 0,
         InitializeParserConfigs,
         RawJsonMerge,
+        WagoDBMerge,
         ContributorDataMerge,
         PreProcessingSetup,
         Validation,
         ConditionalData,
+        Incorporation,
         Consolidation,
         UnsortedGeneration,
         DataIntegrityAnalysis,
