+++ .contrib/Source Code/Parser/packages.config
@@ -1,0 +1,4 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="Csv" version="2.0.93" targetFramework="net48" />
+</packages>
